requirements.txt
Flask
requests
pandas
scikit-learn
numpy
gunicorn
tensorflow-cpu==2.15.1
h5py==3.10.0 
joblib==1.3.2